moves rows_to_json files to data_ingestion folder
